refactor: zoe/src/types-ambient.js
