+++ packages/zoe/exported.js
@@ -4,7 +4,7 @@ import './src/contractFacet/types-ambient.js';
 import './src/zoeService/types.js';
 import './src/contractSupport/types.js';
 import './src/contracts/exported.js';
-import './src/types.js';
+import './src/types-ambient.js';
 import './tools/types-ambient.js';
 import '@agoric/notifier/exported.js';
 import '@agoric/ertp/exported.js';
