+++ packages/zoe/src/zoeService/zoeSeat.js
@@ -1,4 +1,6 @@
 /* eslint @typescript-eslint/no-floating-promises: "warn" */
+import '../types-ambient.js';
+
 import { prepareDurablePublishKit, SubscriberShape } from '@agoric/notifier';
 import { E } from '@endo/eventual-send';
 import { M, prepareExoClassKit } from '@agoric/vat-data';
@@ -6,7 +8,6 @@ import { deeplyFulfilled } from '@endo/marshal';
 import { makePromiseKit } from '@endo/promise-kit';
 
 import { satisfiesWant } from '../contractFacet/offerSafety.js';
-import '../types.js';
 import '../internal-types.js';
 import {
   AmountKeywordRecordShape,
